+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **categories:** Add frequent shortcuts and inline create in the transaction form
+
 ## [0.17.0] - 2026-08-26
 
 ### Added